requirements.txt
ffastapi==0.110.0
uvicorn==0.29.0
pydantic==1.10.14
python-multipart==0.0.9
jinja2==3.1.3
numpy==1.26.4

# For persistent storage
orjson==3.10.6

# For load testing
locust==2.26.0

# For simulations
requests==2.31.0

